% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landforms.R
\name{landforms}
\alias{landforms}
\alias{landforms,SpatRaster,SpatRaster-method}
\alias{landforms,RasterLayer,RasterLayer-method}
\title{Classification of land forms}
\usage{
landforms(small, large, ...)

\S4method{landforms}{SpatRaster,SpatRaster}(small, large, mean_s, mean_l, sd_s, sd_l, ...)
}
\arguments{
\item{small, large}{Raster layers with topographic position index at small
and large scale, respectively.}

\item{...}{Further arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}}.}

\item{mean_s, mean_l}{Reference mean value for small and large scale,
respectively.}

\item{sd_s, sd_l}{Reference standard deviation for small and large scale,
respectively.}
}
\value{
A raster.
}
\description{
Add description.
}
\details{
Details.
}
\examples{
## Add Example

}
\references{
Link to Weiss (2001)
}
