% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_window.R
\name{mask_window}
\alias{mask_window}
\title{Extract portions of windows}
\usage{
mask_window(window, option = "top", dim = c(5, 5))
}
\arguments{
\item{window}{A matrix including a window to be masked, for instance a window generated
by \code{\link[=draw_ring]{draw_ring()}}. If missing, you need to specify the dimensions of the mask.}

\item{option}{A character value indicating the portion to be extracted. The only
alternatives are "top", "bottom", "left", "right", "topleft", "topright",
"bottomleft", and "bottomright".}

\item{dim}{An integer vector of lenght 2. The first value corresponds to the height
(number of rows), and the second value corresponds to the width (number of
columns). Only odd values are accepted. If a 'window' is provided, the dimensions
of this window will be used and this argument will be ignored.}
}
\value{
If window is not provided, a matrix with the specified dimentions and values 1 and
\code{NA}.
If window is provided, It will be masked to the respective portion and excluded cells
will become \code{NA}s.
}
\description{
Some calculations using \code{\link[terra:focal]{terra::focal()}} may require partial windows, for instance
calculating northness or eastness.
Portions may correspond to halves or quarters of the window (see argument \code{option}).
}
\examples{
# Random window with weights
set.seed(42)
window <- matrix(sample(1:3, size = 121, replace = TRUE), ncol = 11)
window

mask_window(option = "right", window = window)
}
\seealso{
\code{\link[=draw_ring]{draw_ring()}}.
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
