% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_sp.R
\name{bbox_sp}
\alias{bbox_sp}
\alias{bbox_sp,Extent-method}
\alias{bbox_sp,RasterLayer-method}
\alias{bbox_sp,RasterStack-method}
\alias{bbox_sp,RasterBrick-method}
\title{Bounding boxes as spatial objects}
\usage{
bbox_sp(raster, ...)

\S4method{bbox_sp}{Extent}(raster, proj4 = "", geom = "point", ...)

\S4method{bbox_sp}{RasterLayer}(raster, geom = "point", ...)

\S4method{bbox_sp}{RasterStack}(raster, geom = "point", ...)

\S4method{bbox_sp}{RasterBrick}(raster, geom = "point", ...)
}
\arguments{
\item{raster}{An object of class \verb{Raster*} or \linkS4class{Extent}.}

\item{...}{Further arguments passed among methods (not yet used).}

\item{proj4}{Character vector indicating the \code{proj4string} (required only
for inputs of class \linkS4class{Extent}).}

\item{geom}{Character value indicating the geometry of output (either point
or polygon).}
}
\value{
Either a \linkS4class{SpatialPointsDataFrame} or a
\linkS4class{SpatialPolygonsDataFrame} object.
}
\description{
Bounding boxes of raster (class \verb{Raster*}) objects will be converted either
to \linkS4class{SpatialPointsDataFrame} or \linkS4class{SpatialPointsDataFrame}.

This function extract the corners including the bounding box of the input
raster or extension and writes either four spatial points or a rectangle as
polygon.
}
\examples{
require(raster)

r <- raster(system.file("external/test.grd", package="raster"))
plot(r)
plot(bbox_sp(r, "polygon"), border="darkgreen", lwd=3, add=TRUE)
plot(bbox_sp(r, "point"), col="red", pch=16, cex=3, add=TRUE)

}
\author{
Lukas Trübenbach and Miguel Alvarez (\email{kamapu78@gmail.com}).
}
