% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodil_raster.R
\name{erodil_raster}
\alias{erodil_raster}
\alias{erodil_raster,RasterLayer-method}
\title{Raster erosion and dilation}
\usage{
erodil_raster(raster, ...)

\S4method{erodil_raster}{RasterLayer}(raster, erosion = TRUE, dilation = TRUE, nt = 1, ...)
}
\arguments{
\item{raster}{An object of class \linkS4class{RasterLayer} containing binary values
(0 or 1).}

\item{...}{Further arguments passed to \code{\link[=raster]{raster()}}.}

\item{erosion}{A logical value indicating whether the data will be eroded or
not.}

\item{dilation}{A logical value indicating whether the data will be dilated
or not.}

\item{nt}{Number of times to be processed (passed to \code{\link[=morphology]{morphology()}}).}
}
\value{
A \linkS4class{RasterLayer} object.
}
\description{
This function apply \code{\link[=morphology]{morphology()}} to raster objects.
}
\author{
Jan Blöthe and Miguel Alvarez (\email{kamapu78@gmail.com}).
}
