% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodil_raster.R
\name{erodil_raster}
\alias{erodil_raster}
\alias{erodil_raster.SpatRaster}
\alias{erodil_raster,SpatRaster-method}
\title{Raster erosion and dilation}
\usage{
erodil_raster(raster, ...)

\method{erodil_raster}{SpatRaster}(
  raster,
  width = c(3, 3),
  type = "diamond",
  erosion = TRUE,
  dilation = TRUE,
  erosion_first = TRUE,
  nt = 1,
  ...
)
}
\arguments{
\item{raster}{An object of class \linkS4class{SpatRaster} containing 1 and 0
(or NAs) values.}

\item{...}{Further arguments passed to \code{\link[=shapeKernel]{shapeKernel()}}.}

\item{width}{An integer vector indicating the width and the heigth applied
for the kernel. This is passed to \code{\link[=shapeKernel]{shapeKernel()}}.}

\item{type}{A character value indicating the type of shape used for the
kernel (see \code{\link[=shapeKernel]{shapeKernel()}})}

\item{erosion}{A logical value indicating whether the data will be eroded or
not.}

\item{dilation}{A logical value indicating whether the data will be dilated
or not.}

\item{erosion_first}{A logical value indicating whether erosion should be
carried out before dilation or vice versa.}

\item{nt}{Number of times to be processed.}
}
\value{
A \linkS4class{SpatRaster} object with value 1 for the processed features and 0
for the background.
The output is already factorized and includes a color table for the two
categories (white and black).
}
\description{
Mixing erosion and dilation processes to clean binary images, getting rid of
isolated pixels and little holes.

For the single steps, see \code{\link[=erode]{erode()}}.
}
\examples{
## Load installed rasterLayer
require(terra)
require(ggplot2)
require(tidyterra)

# Import and prepare data
r <- rast(file.path(path.package("spatialist"), "binras.tif"))
r[is.na(r)] <- 0
r <- as.factor(r)
coltab(r) <- data.frame(value = c(0, 1), col = c("white", "black"))

## Make only erosion or only dilation
r2 <- rast(c(
    original = r,
    eroded = erodil_raster(r, dilation = FALSE),
    dilated = erodil_raster(r, erosion = FALSE)
))

ggplot() + geom_spatraster(data = r2) + facet_wrap(~lyr)

## Erode two times
r2 <- rast(c(
    original = r,
    dilated = erodil_raster(r, erosion = FALSE),
    dilated2 = erodil_raster(r, erosion = FALSE, nt = 2)
))

ggplot() + geom_spatraster(data = r2) + facet_wrap(~lyr)

## By default erosion will be done before dilation
r2 <- rast(c(
  original = r,
  eroded_first = erodil_raster(r),
  dilated_first = erodil_raster(r, erosion_first = FALSE),
  both = erodil_raster(erodil_raster(r, erosion_first = FALSE))
))

ggplot() + geom_spatraster(data = r2) + facet_wrap(~lyr)
}
\author{
Jan Blöthe and Miguel Alvarez (\email{kamapu@posteo.de}).
}
